Factor out mr test utilities.
